+++ clients/ptranslator/src/test/java/org/hawkular/metrics/clients/ptrans/exec/OptionsITest.java
@@ -14,13 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.metrics.clients.ptrans;
+package org.hawkular.metrics.clients.ptrans.exec;
 
 import static org.hamcrest.CoreMatchers.allOf;
-import static org.hawkular.metrics.clients.ptrans.CanReadMatcher.canRead;
-import static org.hawkular.metrics.clients.ptrans.ContainsMatcher.contains;
-import static org.hawkular.metrics.clients.ptrans.HasSizeMatcher.hasSize;
-import static org.hawkular.metrics.clients.ptrans.IsFileMatcher.isFile;
+import static org.hawkular.metrics.clients.ptrans.matchers.CanReadMatcher.canRead;
+import static org.hawkular.metrics.clients.ptrans.matchers.ContainsMatcher.contains;
+import static org.hawkular.metrics.clients.ptrans.matchers.HasSizeMatcher.hasSize;
+import static org.hawkular.metrics.clients.ptrans.matchers.IsFileMatcher.isFile;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertThat;
