+++ clients/ptranslator/src/test/java/org/hawkular/metrics/clients/ptrans/exec/PidFileITest.java
@@ -14,15 +14,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.metrics.clients.ptrans;
+package org.hawkular.metrics.clients.ptrans.exec;
 
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static java.util.concurrent.TimeUnit.SECONDS;
+
 import static org.hamcrest.CoreMatchers.allOf;
-import static org.hawkular.metrics.clients.ptrans.CanReadMatcher.canRead;
-import static org.hawkular.metrics.clients.ptrans.ContainsMatcher.contains;
-import static org.hawkular.metrics.clients.ptrans.IsFileMatcher.isFile;
-import static org.hawkular.metrics.clients.ptrans.WriteLockedMatcher.writeLocked;
+import static org.hawkular.metrics.clients.ptrans.matchers.CanReadMatcher.canRead;
+import static org.hawkular.metrics.clients.ptrans.matchers.ContainsMatcher.contains;
+import static org.hawkular.metrics.clients.ptrans.matchers.IsFileMatcher.isFile;
+import static org.hawkular.metrics.clients.ptrans.matchers.WriteLockedMatcher.writeLocked;
 import static org.hawkular.metrics.clients.ptrans.util.ProcessUtil.getPid;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotEquals;
