+++ clients/ptranslator/src/test/java/org/hawkular/metrics/clients/ptrans/matchers/CanReadMatcher.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.metrics.clients.ptrans;
+package org.hawkular.metrics.clients.ptrans.matchers;
 
 import java.io.File;
 
@@ -26,24 +26,24 @@ import org.hamcrest.TypeSafeMatcher;
 /**
  * @author Thomas Segismont
  */
-public class IsFileMatcher extends TypeSafeMatcher<File> {
+public class CanReadMatcher extends TypeSafeMatcher<File> {
     @Override
     public void describeTo(Description description) {
-        description.appendText("is file");
+        description.appendText("can read");
     }
 
     @Override
     protected void describeMismatchSafely(File item, Description mismatchDescription) {
-        mismatchDescription.appendValue(item).appendText(" is not a file");
+        mismatchDescription.appendValue(item).appendText(" is not readable");
     }
 
     @Override
     protected boolean matchesSafely(File item) {
-        return item.isFile();
+        return item.canRead();
     }
 
     @Factory
-    public static Matcher<File> isFile() {
-        return new IsFileMatcher();
+    public static Matcher<File> canRead() {
+        return new CanReadMatcher();
     }
 }
