+++ clients/ptranslator/src/test/java/org/hawkular/metrics/clients/ptrans/matchers/HasSizeMatcher.java
@@ -14,13 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.metrics.clients.ptrans;
-
-import static org.hawkular.metrics.clients.ptrans.util.Arguments.checkArgument;
+package org.hawkular.metrics.clients.ptrans.matchers;
 
 import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
 
 import org.hamcrest.Description;
 import org.hamcrest.Factory;
@@ -30,35 +26,30 @@ import org.hamcrest.TypeSafeMatcher;
 /**
  * @author Thomas Segismont
  */
-public class ContainsMatcher extends TypeSafeMatcher<File> {
-    private final String content;
+public class HasSizeMatcher extends TypeSafeMatcher<File> {
+    private final long size;
 
-    public ContainsMatcher(String content) {
-        checkArgument(content != null, "content is null");
-        this.content = content;
+    public HasSizeMatcher(long size) {
+        this.size = size;
     }
 
     @Override
     public void describeTo(Description description) {
-        description.appendText("contains");
+        description.appendText("has size");
     }
 
     @Override
     protected void describeMismatchSafely(File item, Description mismatchDescription) {
-        mismatchDescription.appendValue(item).appendText(" has no line with content: ").appendText(content);
+        mismatchDescription.appendValue(item).appendText(" has size: ").appendText(String.valueOf(size));
     }
 
     @Override
     protected boolean matchesSafely(File item) {
-        try {
-            return Files.lines(item.toPath()).anyMatch(line -> line.contains(content));
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
+        return item.length() == size;
     }
 
     @Factory
-    public static Matcher<File> contains(String content) {
-        return new ContainsMatcher(content);
+    public static Matcher<File> hasSize(long size) {
+        return new HasSizeMatcher(size);
     }
 }
