+++ clients/ptranslator/src/test/java/org/hawkular/metrics/clients/ptrans/matchers/WriteLockedMatcher.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.metrics.clients.ptrans;
+package org.hawkular.metrics.clients.ptrans.matchers;
 
 import java.io.File;
 import java.io.RandomAccessFile;
