+++ clients/ptranslator/src/test/java/org/hawkular/metrics/clients/ptrans/util/PrintOutputOnFailureWatcher.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.metrics.clients.ptrans;
+package org.hawkular.metrics.clients.ptrans.util;
 
 import java.io.File;
 import java.io.IOException;
